fixed view matrix, complete spining cube test sample

+++ render.h
@@ -8,6 +8,6 @@
 void clear_terminal(CHAR_T c);
 void flush_terminal();
 
-int set_elem(size_t x, size_t y, CHAR_T c, PREC_T d);
+int set_elem(int x, int y, CHAR_T c, PREC_T d);
 
 int draw_buffer(buff_idx_t id, mat4* vt);
